Flipozon
In this project ,we use
1.jFrame  for display a image of the product.
2.Java file to store a user details.
3.Collection frame work.
4.Packages.
/******************************************/
In this project , we created a path between the sellers and the consumers . In this, first
consumer should give login id number  . If that user not have  account  ,then we create a new account for that user .Else if that user have account ,then we ask  login id ,if given id was matched with our any files then, consumer should  give a type of the product ,then he allow to start search for a particular product of given type, if that product is not  available in the store ,then we ask “do you want to see any other  product in same type…? ” .If consumer accept that request ,then we show the available products in that type, else we ask “do you want to see the products from any other type…?” .If consumer accept the request ,then we show the products from other type. After all purchase, finally we generate the bill . After display the bill,
we created a function for give lucky prize to 
consumers. We ask a user lucky number for give lucky prize ,from that number  we do some calculation, as per result of that calculation give a prize to user.Finally ,we
get a feedback from the consumer.