Demand supply management
A Demand-supply management system will help businessmen/entrepreneurs to find the best location for their next brick and mortar store. The stakeholders who wish to establish an entity queries the system providing information regarding the type of business they wish to start, broader geographical area, target groups, competition and resources. 
The customers/consumers who feel that there is a high demand for a product/service in their location may give the appropriate information like location name, name of the service/product which is in high demand etc as an input to the system helping the stakeholders know that there is a high demand for that service/product in that location. The demand-supply management using the given data by the customers/consumers finds out the best location for the stakeholder.

If a stakeholder wants to establish an entity, more often they face a dilemma about choosing the location as they are unaware whether there is sufficient demand for their product/service in the proposed location. Also, the availability of the necessary materials in the location plays a vital role.
A consumer/customer might feel that there is a huge demand for a product/service in their area but there is no establishment that satisfies the demand. A lot of economic potentials is untapped due to the above- stated problem. Businessmen who try to sell products or provide services to customers in a location where there is no demand for their product/service might face huge financial loss.

By using this system, any stakeholder can finds best location without having separate department for finding desirable and profitable places. So, stakeholder could save  a large amount of money. The stakeholder most of the time end up having insufficient data regarding the demand for their product/service in a location and faces difficulty in locating the best possible location for their entity. We can solve this problem by providing the stakeholder with sufficient data. This data can be collected by conducting a survey or establishing a portal for the customer/consumer for expressing their interest for a particular product/service in their area.

A demand-supply management system will help the stakeholders establish a successful business and increase the quality of life for the consumers/customers whose demands were previously not satisfied. In addition, new jobs will be created.